2025.01.02
캐릭터 선택 및 상점 기능 개선

+++ Assets/Script/AbilityPurchaseData.cs
@@ -5,7 +5,8 @@ public class AbilityPurchaseData : ScriptableObject
 {
     public int[] cost;
     [Space]
-    public int[] additionalAbility;
+    public float[] additionalAbility;
     [Space]
     public int maximumGrade;
+    
 }
